[skylab] Add function to enable querying leases generally in skylab library.
Note, the most general function is intentionally not exported. In general,
we want to make sure that searches for lease tasks are reasonable, and can
be updated when the tags or labels that identify a lease task change in the
future.

BUG=chromium:1069168
TEST=does it compile

Change-Id: Ie77f8b7ff5f94b18a9aac383446724240019135c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2149354
Auto-Submit: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/master@{#30596}