[🚝] SPA: Add permissions Redux duck.
Each resource has a PermissionSet that's an array of Permission strings.
In the Redux store, we keep a mapping of resource name to PermissionSet.

Note that these permissions only apply to the current user.
If the user changes accounts, the permissions will be out of date.
So we'll need to think up a way to refresh the permissions.

Bug: monorail:7543
Change-Id: I9eb5f54ed119ca62173b3b51b0280575d877f9b8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2155176
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30683}