[skylab] Refactor: move inventory v2 client from internal/cmd/inventory to internal/inventory
I want to change the lease-dut task to grab the model from inventory v2
when leasing a DUT by hostname. However, doing this requires me to use
the client that dut-info uses to talk to inventory v2. This means that the
inventory client is now no longer strictly associated with just the inventory
subcommands in skylab and should be moved.

This command is purely a refactor. It should not (and as far as I can tell does not)
introduce any observable changes to the skylab CLI.

The only non-mechanical change I made is changing updateAssets from a free function
to a method in inventoryClientV2.

BUG=chromium:1069168
TEST=make

Change-Id: Ib8167c1b26194c689dc07744e3298c59003f3e0e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2153711
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30732}