[skylab-swarming-worker] Make skylab_swarming_worker target .PHONY
We do not bother to check how recent the go source files are,
therefore we should rebuild skylab-swarming-worker via the makefile
every time the developer asks for it. Technically, the old version
works as well because the skylab_swarming_worker task depends on
the ssw task, which is phony, but that's a bit strange.

BUG=None
TEST=make; make again and confirm that the target is remade

Change-Id: I5e5e13d102e34190b5a510617bbb9612942ae5b6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2150697
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30767}