[🚝] SPA: Update Hotlist Issues page to use User resource names.
The v3 API currently returns a User object for HotlistItem.adder,
but we later decided it should instead be the resource name.
So we're going to migrate by first implementing the front-end logic
using adder.name to fetch the User data. Then we can more easily
switch over the API and front-end simultaneously.

Since the `viewedHotlistIssues` selector already pulls in the Issue
data for all the HotlistItems, we can reuse that logic to also pull
in the User data.

Bug: monorail:7238
Change-Id: Iae94e1c8051ed335755124aeafa5b0a4be74a167
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2159751
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30800}