[skylab] Add implementation of lease-dut for boards.
This CL completes the implementation of lease-dut for boards.
The -board option is an alternative to providing -model or a
hostname. It is also possible to provide additional dimensions
to a board task via the -dims flag.

In order to limit the load on the lab, we are going to limit
the number of per-board leases to 2 per board initially. This
limit may be revised in the future.

BUG=chromium:1069168
TEST=make AND ./skylab lease-dut -board octopus -- verified that it created a swarming task that leased an octopus.

Change-Id: Ic0c907212f9040c2be69fb94d4e59de45f42198a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2160139
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30825}