Add tree_closer config for the infra builders.
This only includes those that are in the existing gatekeeper config.
This set should probably be expanded at some point (for instance, the
only Windows infra builder in the gatekeeper config doesn't exist any
more, so there are no Windows infra-continuous builders here). But for
now I just want to replicate the gatekeeper config to check that
LUCI-Notify does the same as it does.

Bug: 1051371
Change-Id: I40e7b12e29d85799b826ccc088951e510db792de
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2164135
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Commit-Queue: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30954}