[Findit] Findit should also take action if failed build is latest
In

https://source.chromium.org/chromium/infra/infra/+/master:appengine/findit/services/ci_failure.py;l=253

Findit will bail out if the build it is analyzing is the latest build. The function GetLaterBuildsWithAnySameStepFailure only returns non-empty iff there are some builds later than the current build having the same failure.  This means we need to wait for another build to take action on a culprit, which is not the intended behavior.

The intended behaviour should be: take action on culprit, unless there is a later build where the failure no longer appears.


Bug: 1056937
Change-Id: Id3b634e5696d4b0a3cd10ef7a36cf4ffb8886c8b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2167643
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/master@{#31001}