[skylab] Reduce per-model and per-board cap to 1 lease.
This is an immediate-term fix for lease-dut. Some of the models
in the lab, particularly cyan and setzer, have low numbers of
healthy DUTs. We will raise the limit in the near future for
most models, but hardcoding caps for each of the models in the
code itself does not seem like a tenable solution. More details
in bug.

BUG=chromium:1076642
TEST=./skylab lease-dut -board octopus  AND run lease twice and confirm that we hit the limit

Change-Id: I6f0bebe639824361a39edf6aea4139827b5aa2d9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2171830
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/master@{#31049}