Use the id as last_status_key rather than the full key string.
This is what's already used by the JSON API. Using it in the main page
as well will let us fetch the last status from the JSON API, and then
use the key we get from that to update the tree using the main page
endpoint.

This shouldn't cause any backward-compatibility issues, as AFAIK nobody
is hitting this endpoint programmatically. And even if they were, the
only place to obtain the correct key would be by scraping the HTML,
which now contains the key in the new format.

In-flight requests at the moment the update occurs will fail, as they
would get the key in the previous format, but send it to a frontend
expecting keys in the new format. This is okay because firstly it's very
unlikely, as the rate of manual updates is low. And secondly, because
this doesn't cause any problems other than the update failing, and
needing to be re-issued, which will work the second time.

Bug: 1051371
Change-Id: I6919c0b42f2b4ed8e65362a2492788b9d3cab4a5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2192558
Reviewed-by: Joey Scarr <jsca@chromium.org>
Commit-Queue: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#31395}