[skylab] Add -evil-lease flag to lease-dut for emergencies
If we need to do a task such as performing maintenance on
a DUT, then we need to be able to lease it regardless of
per-model limits.

The evil-lease flag is not immediately used when leasing by
hostname because we are not yet enforcing the model cap for
by-hostname leases. Once we start enforcing this limit, we
will use the value of the override flag directly when leasing
by hostname.

This command adds an evil-lease flag and gives it reasonable
semantics. We may want to consider changing the policy in
the future so that, for instance, attempting to lease by
model with the -evil-lease flag passed in results in a
dedicated error message being generated.

BUG=chromium:1069168
TEST=make AND ./skylab lease-dut -model garg -evil-lease AND skylab lease-dut -dev `dut 6 22 13 5`

Change-Id: Iccb0f2b38d604be482f1c6ebba4d2932e3ce805a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2154313
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#31432}