[recipe_roll_tryjob] Rebase onto last non-crashing revision
Before training the downstream repo against the upstream CL, we traverse
backwards from master through the history of the upstream repo until we
find a revision against which training the downstream repo does not
crash. However, until now we didn't actually do anything with that
revision. What we want to do is to rebase the upstream CL on top of that
revision before training to determine triviality.

If the rebase fails, we'll fall back to rebasing on top of upstream
master.

Test: https://ci.chromium.org/swarming/task/4cd82aa8e2e48a10
Change-Id: I9b7a2be418d45c35ce19a55ebc8bc502a120865a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2247228
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Cr-Commit-Position: refs/heads/master@{#32475}