Move chrome-cros-dut builder to a pool without task templates.
This pool recently had task templates added to it, which conflicts
with the packages that BB adds automatically:
https://ci.chromium.org/p/infra/builders/cron/cros-dut-flash/b8877180221531008992

So this moves the builder to a pool that's identical, except it has
no templates.

All this will be obviated once https://crbug.com/1057152 is complete.

Bug: 812428
Change-Id: I04ba622dcf34b2e2f4d942c2817632aae532479c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2252705
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#32526}