Rename template from 'default', fix bug in it.
By calling it "default", we prevent LUCI-Notify from adding its default
email template, and referring to "default" in the email rule ends up
using the status template instead.

The status template has a bug in it, which is fixed here. It worked for
status because status falls back to a default on error. For emails
though, the default is a very spartan template that pretty much just
contains the error message.

Bug: 1095871
Change-Id: I39092c8658bb341550eb84d7e9208226906bd652
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2251640
Auto-Submit: Owen Rodley <orodley@chromium.org>
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Cr-Commit-Position: refs/heads/master@{#32545}