[InventoryV2] Add updated_time when updating registered assets
BUG=chromium:1064071
TEST=Tested on dev

Adds timestamp when updating the registered_assets table. This is just
using the currently empty field to put a timestamp when updating. It is
conceptually not the correct timestamp for updated_time and reflects the
timestamp of the cron job. However, temporarily it can be used to
time partition the registered_assets table.

Change-Id: I786c631c04fc072e2c678b2532351d8dbe5907ca
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2261995
Reviewed-by: Eshwar Narayan <eshwarn@chromium.org>
Commit-Queue: Anushruth  <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#32679}