[stable_version] Use most specific version when validating.
As part of our plan to switch the stable version to use both
the board and model when looking up the CrOS version, we need
to extend the validation tool so that it understands board+model
entries. We do this by picking the most specific possible entry
as the "config version" that's compared against the real entry.

BUG=chromium:1098893

TEST=unittests and ./stable_version2 validate-config -remote-file

Change-Id: I0a2881aab3054f648a0d8886536ed1fd86c8c1ba
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2276182
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#32844}