SoM: Store BQ result in memcache
Currently we have 6 SoM trees that uses the BQ view `sheriff-o-matic.chrome.sheriffable_failures`:
- android
- chromium.gpu.fyi
- chromium
- ios
- chromium.perf
- chrome_browser_release

Each of them has its own cron job to run the analyzer every 5 mins.

From BQ's documentation:
"BigQuery's views are logical views, not materialized views. Because views are not materialized, the query that defines the view is run each time the view is queried. Queries are billed according to the total amount of data in all table fields referenced directly or indirectly by the top-level query"
(There is actually materialized view, which is in Beta, but it has limitation, so we cannot use it)

So it means every 5 mins, we have 6 cron job runs, querying the same view, and in turn, query the underlying tables.

I think we can reduce the cost by caching the result of sheriff-o-matic.chrome.sheriffable_failures, having 1 new cron job to query it every 5 mins and store in memcache, and the 6 current cron jobs just need to read the cache for the data. So we need to filter the result from code side instead of DB side.

As this CL is quite long, I only convert the android tree. Other trees will be handled in the next CL.

Bug: 1092710
Change-Id: Id524ec95a0e189a66c8f1d0b3516eaab05eb0d3b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2275575
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Cr-Commit-Position: refs/heads/master@{#32906}