Delete a bunch of dead Gatekeeper-related code in SOM.
WouldCloseTree is only called from tests. After deleting it, there are
no more users of ClosingSteps or ClosingOptional. The rest of the fields
were dead already.

This makes it more obvious which parts of the config actually need to be
migrated. That is, not very much.

Bug: 1102703
Change-Id: I93e920a819e570d89f221f68e41daa71005370ad
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2283471
Commit-Queue: Owen Rodley <orodley@chromium.org>
Auto-Submit: Owen Rodley <orodley@chromium.org>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33004}