tricium: Remove IsolateWorker as it is now unnecessary
This removes the use of Includes in IsolateWorker;
IsolateWorker was only necessary before because every
analyzer had to have a separate command and the actual
files data was included. Now the input files data
can be used directly for each analyzer as the command
is given in the swarming trigger request.

Bug: 1102545
Change-Id: I6fdadfcdf977dbf3c0aaeb0d88d0e99009418e08
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2287427
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33183}