[dirmeta] Add export subcommand
Rename "extract" to "export", because the latter is a more standard
term for this kind of operation.

Add MappingForm enum for well known forms of mapping.

Subcommand export reads metadata files from a given tree and exports
it all to a single file.
Also it accepts -form flag to indicate which form of the mapping
to export.

Bug: 1104246
Change-Id: I70b05b609ad24c48633cabf0bc17f357bc52cd78
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2292793
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Auto-Submit: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33248}