Monorail: Added dependencies for FakeRedis testing stub
FakeRedis is a library that emulates Redis-py, however, it does
not require a connection to a redis server to simplify running
unit tests.

FakeRedis requires adding a library, sortedcontainers. It also
required an updated version of six.py to use the `ensure_binary`
method.

This CL does the following:
  1. Adds update.sh and README.monorail for FakeRedis
  2. All of the files for FakeRedis to work
  3. update.sh and README.monorail for sortedcontainers
  4. sortedcontainers files
  5. Updated version of six.py

Bug: monorail:7949
Change-Id: I712a07687f9020c5915ff33f199565cd1df922e2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2298140
Commit-Queue: Rob Warner <rwarner@google.com>
Reviewed-by: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33321}