[buildbucket] Add test coverage for repeated dimensions case.
Buildbucket allows to specify repeated dimensions e.g. ["a:1", "a:2"]
which are passed to Swarming (and mean "the bot should have a:1 AND a:2
dimensions").

Add unit tests that cover this case to make sure it really works and
doesn't regress. Fix config validation.

R=nodir@chromium.org
BUG=1105359

Change-Id: I9379390e3d3da0241c475b55f0a3e8054a73f9d0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2299383
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33340}