Manual Repair: implement CreateDeviceManualRepairRecord rpc
CreateDeviceManualRepairRecord takes a repair record message and store
it in the datastore. The hostname in the record is used to find the
asset tag of the DUT in question from the datastore so the user does not
have to enter an asset tag when creating the proto message.

BUG=chromium:1079495
TEST=go test .

Change-Id: I64a1dda9e03afeeb7ac67351ef6ee2d2d835ed95
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2301212
Commit-Queue: Justin Suen <justinsuen@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33377}