Manual Repair: implement UpdateDeviceManualRepairRecord rpc
UpdateDeviceManualRepairRecord takes a repair record message and
attempts to update the existing record in the datastore. It takes the ID
of the Entity and the Record so the updation can be applied to the
correct entity. Timestamps are updated based on RepairState as well.

BUG=chromium:1079495
TEST=go test .

Change-Id: Id88d9c37c1e713dda273518f78fead2383a04ace
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2304150
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/master@{#33402}