Drop preference for dates over distant milestones for expiry
Lots of folks have given feedback that this message is confusing,
because it's not clear where the cutoff is. We could do more to
communicate the cutoff; but now that we have better tooling around
expiry in general, it's also not as important to be able to understand
exactly when an expiry date is set to -- for example, we'll still get a
warning if the chosen milestone is further than 1 year out, and we'll
still have automation to bump the expiry date forward if the metrics are
actively in use.

R=qyearsley

Bug: 1105417
Change-Id: I3995863e56484528b66206b9224f28057071ef66
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2308051
Auto-Submit: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33440}