Monorail: Remove _CheckCompatibility
This function was not used anymore so I removed all uses of it.

'extra_perms_are_sorted' field was removed from the project_pb2
since it was confirmed that removing it was not causing the
profiler_test to fail, as originally thought.

Bug: monorail:7949
Change-Id: I493e47f7ec13f6b748ae3d84caba79bf4b962fca
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2314640
Commit-Queue: Rob Warner <rwarner@google.com>
Auto-Submit: Rob Warner <rwarner@google.com>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33632}