chromeperf: Create top-level directory
This change creates a top-level directory for hosting code being ported
from the `catapult` repository into the infra/infra repository. The
intent is to incrementally move the Python2 code from the Chromeperf
Dashboard and Pinpoint into Python3 code to be deployed as services in
GKE. The first effort that will be included in this directory is the
port of the experimental execution engine into a standalone
implementation.

Creating this directory does not invalidate the chromeperf directory in
`go/src/infra/chromeperf` where all new services that are written in go
will be developed.

R=nodir@chromium.org

Change-Id: Id18d4f800d4acab7fa7091394bca4200e8dd1334
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2331975
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#33801}