Added tagging output 3pp with current infra/3pp revision
Part of the proposal for go/chops-verified-3pp was to tag output cipd
packages with current version of 3pp repository used. This is added
to increase reproducibility, as we would know exactly which revision of
3pp repo was used to build a particular instance of cipd package.

One of the concern was this might cause unnecessary rebuilds of package
when source has not changed, however, before building a package we use
cipd.describe with symver of the package, so if we have built a package
for the symver, we won't be building the same package again.

Bug: 1094120
Change-Id: I46ec23314d7480ecda2478ebc49ab2dbd2942997
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2317094
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Cr-Commit-Position: refs/heads/master@{#33837}