Added feature to check for moving git tags in 3pp recipes.
This feature will catch any moving git tags for 3pp dependencies, for
e.g. if external git repository moves tag v1.2.3 to point to a new git
revision, there is a chance it indicates malicious activity on the
external repo. We would catch such attempts and create warning inside
3pp recipes.

This CL also uploads source checkouts to CIPD as cache in
infra/3pp/sources prefix. The idea here goes beyond source method git,
instead, we want to store the sources locally (local to external parties)
by storing it in our cache the first time we see a new package.

This way we will not fetch sources from outside for further rebuilds of
a package. However, this is not applied in this CL, and will be enabled
at later point of time. This CL only adds functionality to upload
checkout out sources of resolved symver and adds a tag to the source
package for external hash for git method.

Overall idea is to
* return a hash when resolving a symver
* use resolved version and hash to find if package exists in source
cache
* If it does, checks the external_hash tag and matches with current
external hash
* failure in previous stage will result recipe failure, else will
continue normally.

Bug: 1094120
Change-Id: Ib954149dc80dadbac140ab17facdc78298f4c60e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2306738
Reviewed-by: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Cr-Commit-Position: refs/heads/master@{#33949}