[stable_version2] When validating a stable version config file, use cache keyed by board and CrOS version.
Prior to this change, the ./stable_version2 validate-config ... correctly validates a config file
with by-model stable CrOS versions, but fails silently when presented with a file with multiple
distinct CrOS versions within a board. In order to save time, validateconfig.go caches the result
of fetching the metadata.json file for a particular board. However, the cache previously used
only the board name as a key. This means that the cache will incorrectly be used when we try to
download the metadata.json file for a board that has already been used but with a different version.

validated by hand-applying changes associated with the first model in the changeset with change id
I5b2c70d553eb443a08ba9def9ebe3628bd46fc58.

WITH CHANGE:
$ ./stable_version2 validate-config ...
[I2020-08-18T15:11:03.735721-07:00 593610 0 querygs.go:242] creating cache for REDACTED "R77-12371.52.22"
File appears to be a valid stable version config file.

WITHOUT CHANGE:
$ stable_version2 validate-config ...
... details withheld
stable_version2: (2) errors detected

TEST=./stable_version2 validate-config [config file validated by hand] AND ./stable_version2 validate-config -remote-file AND added explicit test case for two models with the same board and different versions
BUG=chromium:1098893

Change-Id: Ic87eacba33611457aaa9c1e5f2d6a83ceecb9f1c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2363908
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/master@{#34426}