Parameterize python for ENV creation.
This adds (and uses) a gclient variable to override which python is
used for the ENV bootstrap creation.

This replaces the previous method where we stuffed the entire system
python installation on $PATH in order to pick up this bootstrap step.

This is needed to unblock monorail folks, as they rely on an actually
functional python (which at least the mac bots don't seem to have)
in order to run pip to install their dependencies.

R=kweng@google.com, tandrii@chromium.org, vadimsh@chromium.org

Change-Id: Ie72e1829dadd1417aa24398eb031f970c619503b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2393224
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Cr-Commit-Position: refs/heads/master@{#34805}