[chromeperf] Port service clients to infra
These service clients have been adapted from code in the Catapult
repository which allow Pinpoint and the Dashboard services to interact
with external services like Gerrit, Gitiles, Isolate, and Monorail.

We need this to be able to support the actions that interact with
external services. We considered using pRPC but at the time of this
writing we learned of a dependency on the NDB tasklet infrastructure,
which we no longer need and are moving away from in the execution
engine.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1116941
Change-Id: Ia17fb5e12cec10a6ba28488636f88066c30436b4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2392229
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Auto-Submit: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#34827}