Use urllib instead of requests in 3pp scripts.
This breaks a circular dependency bootstrapping cpython for a new
platform: .vpython requires the cryptography wheel, which we can't
build until we have a working cpython.

A follow-up step to this will remove requests, cryptography, and the
rest of the dependencies from 3pp/.vpython.

Bug: 1124119
Change-Id: I0546ea1348d0a488d385b156774eae9eccce0396
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2397020
Reviewed-by: Owen Rodley <orodley@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#34888}