[verified-builds] Quick fix to support_3pp failure.
This is the error I was mentioning in my other CL, crrev/c/2345466

It is because we were trying to upload raw source from checkout
directory as a cipd package. But apparently cipd register takes package
spec as input to do the upload.

That's what the above mentioned CL was about. But this fix does a error
handling and instead of error and stopping the further build, it
continues by excepting the error.

Let me know what you think. Another option was to reland the main CL
causing this issue, but that would generate a lot of painful work and
conflicts in the CLs in the chain already.

R=bryner@google.com

Bug: 1094120
Change-Id: I5587f070cff855ac27de3c8fe561ca334114b586
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2399909
Reviewed-by: Brian Ryner <bryner@google.com>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Cr-Commit-Position: refs/heads/master@{#34903}