Reland "Parameterize python for ENV creation."
Parameterize python for ENV creation.

This adds (and uses) a gclient variable to override which python is
used for the ENV bootstrap creation.

This replaces the previous method where we stuffed the entire system
python installation on $PATH in order to pick up this bootstrap step.

This is needed to unblock monorail folks, as they rely on an actually
functional python (which at least the mac bots don't seem to have)
in order to run pip to install their dependencies.

Reland notes:
Same as before, except that build.py now skips rebuilding the ENV,
meaning that CQ shouldn't be broken this time :D.

R=kweng@google.com, tandrii@chromium.org

Bug: monorail:8301
Change-Id: I8d12053ffd37317da6aeecbe7589e0ea6bcc45fb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2399528
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Kevin Weng <kweng@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#34923}