chromeperf: Port find_isolate -> isolate_finder
This change brings in all the required parts of Pinpoint to the infra
repository to allow us to get just the execution engine parts to be
functional. We're not retaining the ndb dependencies and rather are
transitioning to using protocol buffers as the canonical data
representation for state in the engine.

The first evaluator we're porting is the "isolate finder" which gives us
a way to start buildbucket builds and/or use prebuilt isolates.

R=abennetts@google.com, fancl@chromium.org

Change-Id: Ibaf93a00e5f37dc82e2ce985c07ef61c019e756c
Bug: chromium:1116941
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2361986
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Cr-Commit-Position: refs/heads/master@{#34987}