dockerbuild: run git commands from the process cwd rather than the root.
dockerbuild is run via run.py, which already requires your cwd to be in
the repo, so there's no need to specify it explicitly.

This means that the root can now be outside the checkout, which is
useful as it's also used as a temp directory for other things.

Bug: 1110550
Change-Id: Ic07305380c19af1da9703fa39ad1aadd106ac9de
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2413191
Auto-Submit: Owen Rodley <orodley@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#35110}