chromeperf: Port run_test -> test_runner
This change ports the Swarming task runner evaluators from the catapult
repository to infra/infra. It also introduces an example of how an evaluator
can depend on the outputs of another evaluator, by having the dependency
explicitly defined and the inputs checked.

In this change we also do the following:

- Integrate coverage checking when running under `tox`.

- Consolidate the pytest configuration with the tox configuration.

- Refactor a few common actions into mixins (like error reporting).


R=abennetts@google.com, fancl@chromium.org

Change-Id: I0d29b408472adfd9852e05861eb46816c0c1bc12
Bug: chromium:1116941
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2409479
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#35252}