[3pp] Moved extra_tags implementation from spec to upload stage.
Initially, we used to attach extra_tags to cpid package's spec object.
That made sense because it was easy to attach tags to a particular
package spec and carry it over during register phase.

However, while caching the tag for cipd packages, it was creating
confusion because we would attach tags to local cached version, which
can already exist on remote, hence creating a mismatched tag on a
package on local vs remote cipd server.

Most of the usage of tags is closely tied to remote tags and local tags
can be concentrated into a single step, ensure_uploaded, where we can
build and attach extra tags to the package we are registering.

This way, local version of extra_tags will only be built, when remote
server does not have the same version of the package, solving the
discrepancy problem.

In a follow up CL, we will extend the local spec of a package to reflect
remote tags, if a package pre-exists.

R=yiwzhang@google.com

Bug: 1094120
Change-Id: I2babe6c1d5b0b881fa067de3806f898ff0c4ccee
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2423345
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Auto-Submit: Akash Mukherjee <akashmukherjee@google.com>
Cr-Commit-Position: refs/heads/master@{#35276}