[SoM] Fix incorrect build count
There are cases where we cannot get the first time a step failed (e.g. when a step has failed for a long time, or this is a new step with no history)

For the new steps with no history, perhaps we can somehow detect it, but it will require significant changes to the way SoM query BQ. I don't think we want to make the investment now.

For the cases that we don't have the data, we should not show the data, rather than showing the wrong data.

Bug: 1126181
Change-Id: I6d1a451da09fa14131265748dd78de857691de7e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2421991
Auto-Submit: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/master@{#35282}