[3pp] Feature extension for source upload
At it's current state, source_upload is trying to register a dir
as a cipd package which will fail, as cipd register expects a
CIPDSpec instead/package instance.

We have not seen this failure yet, as no new release for third party
packages that use source:git method in last week. When that happens,
3pp builder will fail to build such packages.

This CL fixes that problem by following this steps:
  * talk to the remote to determine current version
  * talk to cipd to see if 'version' is cached
  * compare with remote to see if source hash movement  happened
  * if cipd DOES NOT have a cached version, synthesize it locally,
    then upload it
  * unpack local version

What we did is to create a CIPDSpec when version for a package (to build)
is resolved. This is slightly different from cipd_spec as this is for
sources instead.

Augmented the cipd.build method with a new optional field as
`package_type` indicating whether a package is source type. If so, for
git method, we'd get rid of .git folder and pack it as a cipd package.

Also augmented a new property named `extra_tags` in cipd_spec, this is
very similar to existing `tag` property, but will be only used when
source is present in cipd cache.

This should solve the above stated problem. This CL also considers
expanding and supporting other methods than git, hence built the
supporting extension separately (where necessary) for easy on-boarding
other source methods like `script` or `url`.

Bug: 1094120
Change-Id: I701ed8d5a0c9781bce38f961b24f632ab2e088c9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2345466
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Cr-Commit-Position: refs/heads/master@{#35345}