[SoM] Use SearchIssues instead of BatchGetIssues for annotations
BatchGetIssues is nice, but it returns error when there is a problem with one of the issue requested (e.g. issue not found, issue got deleted...), and in SoM, it is not rare that we got into this (I encountered it 3 times during last week).

From SoM side, there is no way we could know that an issue got deleted.

I think we should use SearchIssues instead. I make the queries run in parallel(8 workers) to reduce the time fetching the issues.

Bug: 1126358
Change-Id: I81491ed50b0c4291b454a642281d9b8df1f97222
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2434028
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/master@{#35472}