weblayer, skew tests: Do check for existing tags in variants.pyl right before cipd search happens
This CL makes a new if statement that checks if a CIPD tag is inside
variants.pyl. If it is not then it will check if the CIPD package
already exists in the CIPD server. If it does not then it will
build and upload a new CIPD package with that tag.

Bug: 1041619, 1099017
Change-Id: Ibb2773e71288699d189709636ad333b235d80837
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2465574
Reviewed-by: Michael Moss <mmoss@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/master@{#35689}