[3pp] Introduced source:url to proto spec for 3pp.
As proposed in go/chops-verified-3pp, we are introducing a new method to
bridge a security gap in source:cipd method. SourceCipd is used when
there's not a stable distribution of third party source code. We
manually download and upload the source artifact to cipd for our 3pp
packages in such instances.

Security risk here is a possibility of tampering the source package
itself. In this new method, we automate the fetch source for source type
cipd and like other methods we upload/cache the source into a cipd
package for future use.

R=jclinton@google.com

Bug: 1099308
Change-Id: I0871fef1a03ebd64854b9ac031d5817572209d9c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2460253
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/master@{#35768}