[cr-buildbucket] Accept both types of pageToken in querying on Build
In Go SearchBuilds, the page token is the id str while in Py, it's a
cursor str generated by datastore in `query_search` and id str in
`tag_index_search`. In order to have a smooth transition from Py to Go,
this CL revises to support both types of page tokens in requests and
only return the id str as the next page token.

Bug: 1090540
Change-Id: I7d4036b00da935ee80a96f3183d0fd53627316bd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2493365
Commit-Queue: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: smut <smut@google.com>
Cr-Commit-Position: refs/heads/master@{#35972}