[findit] Convert SwarmingTaskProperties.command to a list.
It's a list in the Swarming API and always has been:
https://source.chromium.org/chromium/infra/infra/+/master:luci/appengine/swarming/swarming_rpcs.py;l=360;drc=08f7e8215fbfcd7c98f498b76196a6d343a154cd

It appears that this was only working in the past because Swarming
wasn't populating that field.

I've also added code to handle the case where command is populated
instead of extra_args (Swarming will return one or the other).

Bug: 1128541
Change-Id: I1250241dd48fe37c617cd88d01348ca0bf5190e0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2505290
Commit-Queue: Joey Scarr <jsca@chromium.org>
Reviewed-by: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/master@{#36067}