Monorail: Explain when notification emails are due to group membership
Currently, when a group you're a member of gets cced on a bug, all of
the resulting notification emails magically get sent directly to: your
email address. This is pretty cool, but it makes it really hard to
filter unwanted notifications that you get because you're a member of a
big group that often gets mistakenly added to bugs' cc lists (a lot of
the launch approvers lists get mistakenly cced on launch bugs,
chrome-security@ gets unnecessarily cced on a bunch of security bugs,
etc.).

This change adds new text to the footer on emails sent due to group
expansion, to make it easier to filter out these emails.

Fixed: monorail:8602
Change-Id: Iae95c4b56f07c9d9cb2a4ead25d272b07fbe524b
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2510591
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#36149}